task: added function `ExplainSubstrait`
This function calls `translate_mohair` instead of `from_substrait` so
that it can return the physical plan generated from a substrait plan
instead of a physical plan containing only the `from_substrait`
TableFunction

I also added significant style refactoring and re-ordering of some
bindings in `InitializeConnectionMethods`. The re-styling will be
problematic for future rebases, but the readability is worth it for me
until it is upstreamed.