task: updated cmake function to accommodate deps
Now that I added a dependency to the substrait extension, I need to be
able to specify include directories to find headers for the dependency.
The most reliable way I could think of was to updatae the
`build_loadable_extension_directory` function to check for include dirs
at the end of ARGV. This is done by checking if the list of FILES
(source files to build an object file from) has a tail of include
directories. Although hacky, the approach is to iteratively check the
last entry in FILES to see if it is prefixed with '-I' and if so, peel
it off. Otherwise, leave it.

This approach should allow for existing usage to remain unchanged.