task: update some cmake logic for extensions
This update is to use a similar and compatible approach as what we can
implement with meson: (1) similar approach is to accumulate generated
content into a variable and then to use template substitution
(configure_file) to create a generated file, (2) compatible approach
means using variable syntax in template files (*.in) that both cmake and
meson recognize as variables (@<variable_name>@).

Oddly, the cmake build used 2 different approaches to generate files,
the .cpp file was generated using template substitution; whereas, the
header file was generated using cmake's `file(APPEND...)` function.
Meson supports very similar template substitution for configuration
files, so we change the cmake build to only use this method.

The template files (*.in) originally used `${<variable_name>}` syntax
for specifying variables, but both cmake and meson support
`@<variable_name>@` syntax, so we changed the syntax in the template
files.